+++ genium/src/test/resources/smoke/structs_with_companion/output/android/jni/com_example_smoke_StructsWithMethodsInterface_Vector3.cpp
@@ -77,7 +77,7 @@ Java_com_example_smoke_StructsWithMethodsInterface_00024Vector3_create__Lcom_exa
         _jenv->Throw(static_cast<jthrowable>(exception.release()));
         return nullptr;
     }
-    auto result = nativeCallResult.safe_value();
+    auto result = nativeCallResult.unsafe_value();
     return ::genium::jni::convert_to_jni(_jenv, result).release();
 }
 }