+++ genium/src/test/resources/smoke/structs_with_companion/output/android/jni/com_example_smoke_Vector.cpp
@@ -74,7 +74,7 @@ Java_com_example_smoke_Vector_create__Lcom_example_smoke_Vector_2(JNIEnv* _jenv,
         _jenv->Throw(static_cast<jthrowable>(exception.release()));
         return nullptr;
     }
-    auto result = nativeCallResult.safe_value();
+    auto result = nativeCallResult.unsafe_value();
     return ::genium::jni::convert_to_jni(_jenv, result).release();
 }
 }