+++ genium/src/test/resources/smoke/structs_with_companion/output/cbridge/src/smoke/cbridge_StructsWithMethods.cpp
@@ -64,7 +64,7 @@ _baseRef smoke_StructsWithMethods_Vector_create_xy(double x, double y) {
 smoke_StructsWithMethods_Vector_create_copy_result smoke_StructsWithMethods_Vector_create_copy(_baseRef other) {
     auto&& RESULT = ::smoke::Vector::create(Conversion<::smoke::Vector>::toCpp(other));
     if (RESULT.has_value()) {
-        return {true, .returned_value = Conversion<::smoke::Vector>::toBaseRef(RESULT.safe_value())
+        return {true, .returned_value = Conversion<::smoke::Vector>::toBaseRef(RESULT.unsafe_value())
 };
     } else {
         return {false, .error_code = static_cast< smoke_ValidationUtils_ValidationError >(RESULT.error().value())};